\name{nice_hist}
\alias{nice_hist}
\title{
Histogram Function
}
\description{
This function creates a nice histogram with a given data input.
}
\usage{
nice_hist(data, var,
                bins = 15,
                title = "Histogram",
                x_label = NULL,
                y_label = "Count"))
}
\arguments{
  \item{data}{input data}
  \item{var}{variable}
  \item{bins}{binwidth}
  \item{title}{title of histogram}
  \item{x_label}{x label}
  \item{y_label}{y label
}
}

\author{
Rahel Hintermann
}
