\name{nice_point_plot}
\alias{nice_point_plot}
\title{
Plot Function
}
\description{
This function creates a nice plot with a given data input.
}
\usage{
nice_point_plot(data, x, y, title)
}
\arguments{
  \item{data}{input data}
  \item{x}{independent variable}
  \item{y}{dependent variable}
  \item{title}{title of histogram}
}

\author{
Rahel Hintermann
}

