% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlots.R
\name{niceBinscatter}
\alias{niceBinscatter}
\title{Binscatter}
\usage{
niceBinscatter(
  data,
  x,
  y,
  color = NULL,
  type = "pointrange",
  bins = 20,
  xlims = NULL,
  ylims = NULL,
  title = "",
  xlab = x,
  ylab = y,
  colorname = color,
  colorlabels = NULL,
  legendpos = "right"
)
}
\arguments{
\item{data}{A dataset}

\item{x}{horizontal axis variable}

\item{y}{vertical axis variable}

\item{color}{Character. Variables to group on. With changing color}

\item{type}{Type of geom to plot ("point" (default), "pointrange", "line")}

\item{bins}{Numeric. Number of bins (default: 20)}

\item{xlims, ylims}{Numeric vectors of 2 elements. Axis limits}

\item{title}{Plot title}

\item{xlab, ylab}{Axis labels}

\item{colorname}{Optional variable names for groups to be displayed on legend}

\item{colorlabels}{Optional labels for grouos to be displayed on legend}

\item{legendpos}{Legend position. Options: “left”,“top”, “right” (default), “bottom”}
}
\value{
A ggplot2-plot
}
\description{
Group variable on the horizontal axis into equal-sized bins and calculate group means
for each bin.
}
\details{
Backend code source: https://github.com/maximilianeber/binscatter (Maximilian Eber)
}
\examples{
niceBinscatter(ggplot2::mpg, "displ", "hwy", color = "class", type = "pointrange")
}
