% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlots.R
\name{niceDensity}
\alias{niceDensity}
\title{Nice density plot}
\usage{
niceDensity(
  data,
  x,
  fill = NULL,
  weight = 1,
  color = NULL,
  alpha = 0.3,
  bw = NULL,
  xlims = NULL,
  ylims = NULL,
  title = "",
  xlab = x,
  ylab = "density",
  colorname = color,
  fillname = fill,
  colorlabels = NULL,
  filllabels = NULL,
  legendpos = "right"
)
}
\arguments{
\item{data}{A dataset}

\item{x}{horizontal axis variable}

\item{weight}{Character. Optional variable with weights.}

\item{color, fill}{Character. Variables to group on. With changing fill or border color}

\item{alpha}{Opacity of fill color}

\item{bw}{Numeric. Bandwidth. Controls the smoothness of the plot}

\item{xlims, ylims}{Numeric vectors of 2 elements. Axis limits}

\item{title}{Plot title}

\item{xlab, ylab}{Axis labels. Ylab default: "density"}

\item{fillname, colorname}{Optional variable names for groups to be displayed on legend}

\item{filllabels, colorlabels}{Optional labels for grouos to be displayed on legend}

\item{legendpos}{Legend position. Options: “left”,“top”, “right” (default), “bottom”}
}
\value{
A ggplot2-plot
}
\description{
Nice density plot
}
\examples{
niceDensity(ggplot2::mpg, "displ", bw = 0.2, color = "class", fill = "class")
}
