% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlots.R
\name{niceDiscreteLabels}
\alias{niceDiscreteLabels}
\title{Change the labels of discrete scales manually}
\usage{
niceDiscreteLabels(
  p,
  linename = NULL,
  linetypelabels = NULL,
  colorname = NULL,
  colorlabels = NULL,
  fillname = NULL,
  filllabels = NULL
)
}
\arguments{
\item{p}{A ggplot2-plot}

\item{linename, colorname, fillname}{The names of the scales}

\item{linetypelabels, colorlabels, filllabels}{Vectors of labels, one for each}
}
\value{
A ggplot2-plot
}
\description{
Implemented as part of nicePlot functions. Not meant for direct use.
}
