% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlots.R
\name{niceHistogram}
\alias{niceHistogram}
\title{Nice histogram}
\usage{
niceHistogram(
  data,
  x,
  fill = NULL,
  color = NULL,
  alpha = 0.3,
  binwidth = NULL,
  xlims = NULL,
  ylims = NULL,
  title = "",
  xlab = x,
  ylab = "density",
  colorname = color,
  fillname = fill,
  colorlabels = NULL,
  filllabels = NULL,
  legendpos = "right"
)
}
\arguments{
\item{data}{A dataset}

\item{x}{horizontal axis variable}

\item{color, fill}{Character. Variables to group on. With changing fill or border color}

\item{alpha}{Opacity of fill color}

\item{binwidth}{Numeric. Width of bins}

\item{xlims, ylims}{Numeric vectors of 2 elements. Axis limits}

\item{title}{Plot title}

\item{xlab, ylab}{Axis labels. Ylab default: "density"}

\item{fillname, colorname}{Optional variable names for groups to be displayed on legend}

\item{filllabels, colorlabels}{Optional labels for grouos to be displayed on legend}

\item{legendpos}{Legend position. Options: “left”,“top”, “right” (default), “bottom”}
}
\value{
A ggplot2-plot
}
\description{
Nice histogram
}
\examples{
niceHistogram(ggplot2::mpg, "displ", binwidth = 0.3, color = "class", fill = "class")
}
