% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlots.R
\name{niceLinePlot}
\alias{niceLinePlot}
\title{Nice line plot, possibly with many groups}
\usage{
niceLinePlot(
  data,
  x,
  y,
  group = NULL,
  color = NULL,
  linetype = NULL,
  line_size = 0.7,
  xlims = NULL,
  ylims = NULL,
  title = "",
  xlab = x,
  ylab = y,
  linename = linetype,
  colorname = color,
  linetypelabels = NULL,
  colorlabels = NULL,
  legendpos = "right"
)
}
\arguments{
\item{data}{A dataset}

\item{x}{horizontal axis variable}

\item{y}{vertical axis variable}

\item{group, color, linetype}{Character. Variables to group on. With changing linetype or color}

\item{line_size}{Numeric. General size of lines}

\item{xlims, ylims}{Numeric vectors of 2 elements. Axis limits}

\item{title}{Plot title}

\item{xlab, ylab}{Axis labels}

\item{linename, colorname}{Optional variable names for groups to be displayed on legend}

\item{linetypelabels, colorlabels}{Optional labels for grouos to be displayed on legend}

\item{legendpos}{Legend position. Options: “left”,“top”, “right” (default), “bottom”}
}
\value{
A ggplot2-plot
}
\description{
Nice line plot, possibly with many groups
}
\examples{
niceLinePlot(ggplot2::mpg, "displ", "hwy", color = "class", linetype = "class",  ylim = c(10,50), line_size = 1)

# Changes to the label variables can be done directly
niceLinePlot(datasets::mtcars, "disp", "hp", color = "vs", colorlabels = c("V-shaped", "straight"), colorname = "Engine")
}
