% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_binscatter.R
\name{stat_binscatter}
\alias{stat_binscatter}
\title{Binscatter}
\usage{
stat_binscatter(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{geom}{Which geom to plot (defaults to "point", other options include "pointrange" or
"line")}

\item{bins}{Number of bins (defaults to 10)}
}
\description{
Group variable on the horizontal axis into equal-sized bins and calculate group means
for each bin
}
